Title: 십진수와 이진수, Time: -1ms, Memory: -1MB, Status: Runtime Error - CodeTree

+++ 240519/십진수와 이진수/decimal-and-binary-number.java
@@ -5,9 +5,9 @@ public class Main {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 
         String N = br.readLine();
-        int num = Integer.parseInt(N, 2) * 17;
+        long num = Long.parseLong(N, 2) * 17;
 
-        String answer = Integer.toBinaryString(num);
+        String answer = Long.toBinaryString(num);
         System.out.println(answer);
     }
 }
